cao: Connection Attempt Object
Wrap struct lws_conmon into an lws_cao_t, and give that its own lifecycle
distinct from any wsi.

Conmon contains what we want, but it was composed into the wsi and so had
to follow its single wsi's lifecycle.

Let's give it its own lifecycle independent of the wsi it happened to be
born for, and wrap it with an lws_cao_t that can contain private things.
We can still present the lws_conmon_t unchanged to the public apis that
expect it.

Client wsi with the connection have at least one CAO, the HEAD CAO is taken
as the active one for a wsi, it may acquire other "losing" CAOs as it goes
along.

desc moves into this new CAO, it's a bit messy because only client
connections have CAOs, but server connections and listeners have desc.  So
we have a .desc in the wsi still for that, and desc in the CAO for client,
and new helpers to take care of which to use.